Updated dicm2nii
and cleaned SPGR_fit code

+++ Common/FitData.m
@@ -48,9 +48,9 @@ Fit.computed = Fit.(fields{1});
 
 % Apply mask
 if (~isempty(data.Mask))
-    Fit.Mask = double(data.Mask);
+    Fit.Mask = single(data.Mask);
     Mask = reshape(data.Mask,nV,1);
-    MTdata = MTdata.*repmat(single(Mask),[1,nT]);
+    MTdata = MTdata.*repmat(Mask,[1,nT]);
 end
 
 % Find voxels that are not empty
